Big project changes related to HW-04 and HW-05: 1. Create separate module based on common part used in both of the HWs 2. Implement most of changes requested in review 3. Update code style and improve readability
