--- homework-04/src/main/java/ru/mail/track/ui11/hw04/navigation/PageUrl.java
@@ -1,12 +1,0 @@
-package ru.mail.track.ui11.hw04.navigation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface PageUrl {
-    String value() default "";
-}
