--- homework-04/src/test/java/ru/mail/track/ui11/hw04/NewsPetsMailPageTest.java
@@ -1,36 +1,0 @@
-package ru.mail.track.ui11.hw04;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.WebDriver;
-import ru.mail.track.ui11.hw04.data.BrowserFactory;
-import ru.mail.track.ui11.hw04.drivers.WebDriverFactory;
-import ru.mail.track.ui11.hw04.pages.NewsPetsMailPage;
-
-public class NewsPetsMailPageTest {
-
-    private WebDriver driver = null;
-
-    @Before
-    public void init() {
-        driver = WebDriverFactory.getWebDriverInstance(BrowserFactory.getBrowser(System.getProperty("webdriver.driver")));
-    }
-
-    @After
-    public void killSession() {
-        driver.close();
-        driver.quit();
-    }
-
-    @Test
-    public void check_show_more_button(){
-        new NewsPetsMailPage(driver)
-                .open()
-                .pageShallBeOpened()
-                .checkNumberOfNewsInList()
-                .pressShowMoreButton()
-                .checkNumberOfNewsInListAfterButtonPressing()
-                .checkNewsAfterButtonPressing();
-    }
-}
