+++ homework-05/src/main/java/ru/mail/track/ui11/hw05/task01/PogodaMailPage.java
@@ -6,14 +6,14 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import ru.mail.track.ui11.hw04.navigation.Domain;
-import ru.mail.track.ui11.hw04.navigation.PageUrl;
-import ru.mail.track.ui11.hw04.navigation.UrlPattern;
-import ru.mail.track.ui11.hw04.pages.AbstractPage;
+import ru.mail.track.ui11.seleniumtestcore.navigation.Domain;
+import ru.mail.track.ui11.seleniumtestcore.navigation.PageUrl;
+import ru.mail.track.ui11.seleniumtestcore.navigation.UrlPattern;
+import ru.mail.track.ui11.seleniumtestcore.AbstractPage;
 
 import static org.junit.Assert.assertTrue;
 
-@UrlPattern("/prognoz/[\\w]{1,20}/")
+@UrlPattern("/prognoz/[\\w_\\-]+/")
 @PageUrl("/")
 @Domain("https://pogoda.mail.ru")
 public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
@@ -24,13 +24,7 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
     @FindBy(css = ".pm-toolbar__search__input")
     private WebElement searchInputField;
 
-    @FindBy(css = "span[class=\"pm-toolbar__suggests__group__item  pm-toolbar__suggests__group__item_with-caption\"]:first-child > span > span > span:first-child")
-    private WebElement cityElement;
-
-    @FindBy(xpath = "//div[@class='information__header__left__place']/h1[text()='Киев']")
-    private WebElement cityOnPage;
-
-    @FindBy(css = "span[class=\"icon icon_fav-flag js-toggle_fav\"]")
+    @FindBy(xpath = "//span[contains(@title, 'Добавить в избранные')]")
     private WebElement addToFavoritesElement;
 
     private final String cityDropdownSelector = "[data-class='pm-dropdown']";
@@ -42,6 +36,14 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
         super(driver);
     }
 
+    public PogodaMailPage cityDropdownShallBeNotPresent() {
+        assertTrue("Выпадающее меню города отображается на странице",
+                standardWaiter.waitForCondition(ExpectedConditions.not(ExpectedConditions
+                        .attributeContains(By.cssSelector(cityDropdownSelector),
+                                "class", "toolbar__dropdown_show"))));
+        return this;
+    }
+
     public PogodaMailPage moveCursorToCityButton() {
         new Actions(driver)
                 .moveToElement(toolbarCityButton)
@@ -50,11 +52,11 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
         return this;
     }
 
-    public PogodaMailPage cityDropdownShouldBeNotPresent() {
-        assertTrue("Выпадающее меню города отображается на странице",
-                standardWaiter.waitForCondition(ExpectedConditions.not(ExpectedConditions
+    public PogodaMailPage cityDropdownShallBePresent() {
+        assertTrue("Выпадающее меню города не отображается на странице",
+                standardWaiter.waitForCondition(ExpectedConditions
                         .attributeContains(By.cssSelector(cityDropdownSelector),
-                                "class", "toolbar__dropdown_show"))));
+                                "class", "toolbar__dropdown_show")));
         return this;
     }
 
@@ -64,8 +66,7 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
         return this;
     }
 
-
-    public PogodaMailPage suggestShouldBeEnabled() {
+    public PogodaMailPage suggestShallBeEnabled() {
         assertTrue("Саджесты отображаются на странице",
                 standardWaiter.waitForCondition(ExpectedConditions.not(ExpectedConditions
                         .attributeContains(By.cssSelector(cityDropdownSelector),
@@ -73,9 +74,11 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
         return this;
     }
 
-    public PogodaMailPage pressCity() {
+    public PogodaMailPage pressFirstSuggestedCity() {
+        String firstSuggestedCity = String.format("//span[@class='pm-toolbar__suggests__group__item  pm-toolbar__suggests__group__item_with-caption']//span[text()='%s']", city);
+        WebElement firstSuggestedCityElement = driver.findElement(By.xpath(firstSuggestedCity));
         new Actions(driver)
-                .moveToElement(cityElement)
+                .moveToElement(firstSuggestedCityElement)
                 .click()
                 .build()
                 .perform();
@@ -85,7 +88,7 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
     public PogodaMailPage pageWithForecastShouldBeOpen() {
         assertTrue("Город отображается на странице",
                 standardWaiter.waitForCondition((
-                        ExpectedConditions.visibilityOf(cityOnPage))));
+                        ExpectedConditions.visibilityOf(driver.findElement(By.xpath(String.format("//div[@class='information__header__left__place']/h1[text()='%s']", city)))))));
         return this;
     }
 
@@ -98,8 +101,8 @@ public class PogodaMailPage extends AbstractPage<PogodaMailPage> {
         return this;
     }
 
-    public PogodaMailPage cityShouldBePresented() {
-        assertTrue("Должен быть выбран город \"" + city + "\"",
+    public PogodaMailPage cityShouldBePresentedInDropdown() {
+        assertTrue(String.format("Должен быть выбран город \"%s\"", city),
                 standardWaiter.waitForCondition(ExpectedConditions
                         .textToBe(By.xpath(cityInFavorites), city)));
         return this;
