+++ homework-05/src/main/java/ru/mail/track/ui11/hw05/task02/NavigateButton.java
@@ -1,0 +1,17 @@
+package ru.mail.track.ui11.hw05.task02;
+
+public enum NavigateButton {
+
+    PREV("prev"),
+    NEXT("next");
+
+    private final String name;
+
+    NavigateButton(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
