+++ homework-05/src/main/java/ru/mail/track/ui11/hw05/task04/NewsSportMailPage.java
@@ -3,15 +3,15 @@ package ru.mail.track.ui11.hw05.task04;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import ru.mail.track.ui11.hw04.navigation.Domain;
-import ru.mail.track.ui11.hw04.navigation.PageUrl;
-import ru.mail.track.ui11.hw04.navigation.UrlPattern;
-import ru.mail.track.ui11.hw04.pages.AbstractPage;
+import ru.mail.track.ui11.seleniumtestcore.navigation.Domain;
+import ru.mail.track.ui11.seleniumtestcore.navigation.PageUrl;
+import ru.mail.track.ui11.seleniumtestcore.navigation.UrlPattern;
+import ru.mail.track.ui11.seleniumtestcore.AbstractPage;
 
 import static org.junit.Assert.assertTrue;
 import static org.openqa.selenium.support.ui.ExpectedConditions.not;
 
-@UrlPattern("/news/football-worldcup/[\\d]{8}/")
+@UrlPattern("/[\\w_\\-]+/[\\w_\\-]+/[\\d]{8}/")
 @PageUrl("/news/football-worldcup/32246552/")
 @Domain("https://sportmail.ru")
 public class NewsSportMailPage extends AbstractPage<NewsSportMailPage> {
@@ -28,4 +28,5 @@ public class NewsSportMailPage extends AbstractPage<NewsSportMailPage> {
                         not(ExpectedConditions.presenceOfAllElementsLocatedBy(By.xpath(error)))));
         return this;
     }
+
 }
