+++ homework-05/src/test/java/ru/mail/track/ui11/hw05/task03/CarsMailPageTest.java
@@ -4,8 +4,8 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
-import ru.mail.track.ui11.hw04.data.BrowserFactory;
-import ru.mail.track.ui11.hw04.drivers.WebDriverFactory;
+import ru.mail.track.ui11.seleniumtestcore.data.BrowserFactory;
+import ru.mail.track.ui11.seleniumtestcore.drivers.WebDriverFactory;
 
 public class CarsMailPageTest {
 
@@ -25,8 +25,7 @@ public class CarsMailPageTest {
     @Test
     public void open_and_close_cars_mail_popup() {
         new CarsMailPage(driver)
-                .open()
-                .pageShallBeOpened()
+                .open("catalog", "ford", "focus", "iii_restailing", "sedan")
                 .clickFirstPosAssessment()
                 .popupShallBePresent()
                 .checkPopupTitle()
