+++ pom.xml
@@ -14,6 +14,7 @@
         <module>homework-03</module>
         <module>homework-04</module>
         <module>homework-05</module>
+        <module>selenium-test-core</module>
     </modules>
     <packaging>pom</packaging>
