+++ selenium-test-core/src/main/java/ru/mail/track/ui11/seleniumtestcore/data/BrowserFactory.java
@@ -1,0 +1,14 @@
+package ru.mail.track.ui11.seleniumtestcore.data;
+
+public class BrowserFactory {
+
+    public static BrowsersData getBrowser(String browser) {
+
+        switch (browser) {
+            case "chrome":
+                return BrowsersData.Chrome;
+            default:
+                return null;
+        }
+    }
+}
