+++ selenium-test-core/src/main/java/ru/mail/track/ui11/seleniumtestcore/data/BrowsersData.java
@@ -1,0 +1,16 @@
+package ru.mail.track.ui11.seleniumtestcore.data;
+
+public enum BrowsersData {
+
+    Chrome("Google Chrome");
+
+    private final String name;
+
+    BrowsersData(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
