+++ selenium-test-core/src/main/java/ru/mail/track/ui11/seleniumtestcore/navigation/Domain.java
@@ -1,0 +1,12 @@
+package ru.mail.track.ui11.seleniumtestcore.navigation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Domain {
+    String value();
+}
